feat: Getting started screen with user account validation via email
Implemented the getting started screen which asks ueser for the email to validate whether the
user already has an account or not.
if the user has an account already the starting screen will navigate directly to the login screen
after the email validation. if user doesn't have an account it will redirect user to the signup screen.

Implemented #20